changed countIntLength type from I to UL

+++ include/misc.h
@@ -8,6 +8,6 @@ void generateTextKeyChain(char *key, int *textKeyChain, int textKeyChainSize); /
 void deallocateOutput(outputFlags flags); //DEALLOCATES flags
 void checkKey(char *key, inputFlags flags); //CHECKS IF KEY IS VALID
 void checkText(char *text, inputFlags flags); //CHECKS IF TEXT IS VALID
-int countIntLength(int number); //RETURNS LENGTH OF number
+unsigned long countIntLength(int number); //RETURNS LENGTH OF number
 
 #endif
